fix: renames pages dir to templates
- renames pages dir to templates, to remove it from next build
